Allows for multiple scenarios to display on server
This was the final result of the weekend Hackathon in Copenhagen (June 2018) for displaying the web pages of the uncertainty and entropy cubes. It reads in pickles of the formatted data, and displays them interactively on the server. It adds the horizons of each realization. It depends on the template of html. It also depends on numerous large pickle files that were too big (?) to upload to Github.

Alas, we never fully achieved the idea of interactively selecting between scenarios. Right now it is fixed in the URL. This is the obvious next idea for enhancement.